requirements.txt
openai==0.25.0 
twilio==7.16.0
flask==0.0.25
python-dotenv==0.21.0
pyngrok==5.2.1
flask-ngrok==0.0.25
deep-translator==1.9.1